refactor(vize): split check/runner into cohesive modules
