+++ crates/vize/src/commands/check/runner/tests.rs
@@ -1,0 +1,342 @@
+use super::{
+    collect_project_global_component_stubs, find_nearest_tsconfig_dir,
+    is_suppressed_false_positive, resolve_declaration_dir, resolve_declaration_emit_options,
+    resolve_project_root, resolve_tsconfig_path, validate_corsa_server_count,
+    write_nuxt_fallback_tsconfig,
+};
+use crate::commands::check::nuxt::NuxtPathAlias;
+use crate::commands::check::tsconfig_inputs::TsconfigDeclarationOptions;
+use std::{
+    path::{Path, PathBuf},
+    sync::atomic::{AtomicUsize, Ordering},
+};
+
+fn unique_case_dir(name: &str) -> PathBuf {
+    static NEXT_CASE_ID: AtomicUsize = AtomicUsize::new(0);
+
+    let case_id = NEXT_CASE_ID.fetch_add(1, Ordering::Relaxed);
+    PathBuf::from(env!("CARGO_MANIFEST_DIR"))
+        .join("target")
+        .join("vize-tests")
+        .join(format!(
+            "check-runner-{name}-{}-{case_id}",
+            std::process::id()
+        ))
+}
+
+#[test]
+fn suppresses_nuxt_nitro_import_meta_conflict_false_positive() {
+    let diagnostic = vize_canon::BatchDiagnostic {
+        file: PathBuf::from("app/app.vue"),
+        line: 0,
+        column: 0,
+        message: "Interface 'ImportMeta' cannot simultaneously extend types 'NitroStaticBuildFlags' and 'NitroImportMeta'.\nNamed property 'preset' of types 'NitroStaticBuildFlags' and 'NitroImportMeta' are not identical.".into(),
+        code: Some(2320),
+        severity: 1,
+        block_type: None,
+    };
+
+    assert!(is_suppressed_false_positive(&diagnostic));
+
+    let mut unrelated = diagnostic.clone();
+    unrelated.message = "Interface 'Other' cannot simultaneously extend types".into();
+    assert!(!is_suppressed_false_positive(&unrelated));
+}
+
+#[test]
+fn collects_project_global_component_stubs_from_ambient_dts() {
+    let project_root = unique_case_dir("global-components");
+    let _ = std::fs::remove_dir_all(&project_root);
+    let src_dir = project_root.join("src");
+    std::fs::create_dir_all(&src_dir).unwrap();
+    let dts_path = src_dir.join("components.d.ts");
+    std::fs::write(
+        &dts_path,
+        r#"import "vue";
+declare module "vue" {
+  export interface GlobalComponents {
+    GlobalComponent: typeof import("./GlobalComponent.vue")["default"]
+  }
+}
+export {};
+"#,
+    )
+    .unwrap();
+
+    let mut options = vize_canon::virtual_ts::VirtualTsOptions::default();
+    collect_project_global_component_stubs(
+        &mut options,
+        std::slice::from_ref(&dts_path),
+        &project_root,
+        None,
+    );
+
+    assert_eq!(options.external_template_bindings, ["GlobalComponent"]);
+    assert!(
+        options.auto_import_stubs.iter().any(|stub| {
+            stub.contains("declare const GlobalComponent:")
+                && stub.contains("./src/GlobalComponent.vue.ts")
+        }),
+        "missing GlobalComponent stub: {:?}",
+        options.auto_import_stubs
+    );
+
+    let _ = std::fs::remove_dir_all(&project_root);
+}
+
+#[test]
+fn resolves_monorepo_root_for_files_spanning_package_tsconfigs() {
+    let project_root = unique_case_dir("monorepo-root");
+    let _ = std::fs::remove_dir_all(&project_root);
+    let app_dir = project_root.join("packages/app");
+    let ui_dir = project_root.join("packages/ui");
+    std::fs::create_dir_all(app_dir.join("src")).unwrap();
+    std::fs::create_dir_all(ui_dir.join("src")).unwrap();
+    std::fs::write(project_root.join("tsconfig.json"), "{}").unwrap();
+    std::fs::write(app_dir.join("tsconfig.json"), "{}").unwrap();
+    std::fs::write(ui_dir.join("tsconfig.json"), "{}").unwrap();
+    let files = vec![app_dir.join("src/App.vue"), ui_dir.join("src/UiButton.vue")];
+    for file in &files {
+        std::fs::write(file, "<template />").unwrap();
+    }
+
+    let resolved_root = resolve_project_root(None, &project_root, &files);
+    let resolved_tsconfig = resolve_tsconfig_path(None, &project_root, &resolved_root, &files);
+
+    assert_eq!(resolved_root, project_root);
+    assert_eq!(resolved_tsconfig, Some(resolved_root.join("tsconfig.json")));
+
+    let _ = std::fs::remove_dir_all(&resolved_root);
+}
+
+#[test]
+fn resolves_package_root_for_files_inside_one_package() {
+    let project_root = unique_case_dir("package-root");
+    let _ = std::fs::remove_dir_all(&project_root);
+    let app_dir = project_root.join("packages/app");
+    std::fs::create_dir_all(app_dir.join("src")).unwrap();
+    std::fs::write(project_root.join("tsconfig.json"), "{}").unwrap();
+    std::fs::write(app_dir.join("tsconfig.json"), "{}").unwrap();
+    let files = vec![app_dir.join("src/App.vue"), app_dir.join("src/main.ts")];
+    for file in &files {
+        std::fs::write(file, "").unwrap();
+    }
+
+    let resolved_root = resolve_project_root(None, &project_root, &files);
+    let resolved_tsconfig = resolve_tsconfig_path(None, &project_root, &resolved_root, &files);
+
+    assert_eq!(resolved_root, app_dir);
+    assert_eq!(resolved_tsconfig, Some(resolved_root.join("tsconfig.json")));
+
+    let _ = std::fs::remove_dir_all(&project_root);
+}
+
+#[test]
+fn resolves_common_root_when_explicit_tsconfig_is_below_inputs() {
+    let project_root = unique_case_dir("explicit-tsconfig-below-inputs");
+    let _ = std::fs::remove_dir_all(&project_root);
+    let config_dir = project_root.join("config");
+    let src_dir = project_root.join("src");
+    std::fs::create_dir_all(&config_dir).unwrap();
+    std::fs::create_dir_all(&src_dir).unwrap();
+    let tsconfig = config_dir.join("tsconfig.json");
+    let app = src_dir.join("App.vue");
+    std::fs::write(&tsconfig, "{}").unwrap();
+    std::fs::write(&app, "<template />").unwrap();
+    let files = vec![app];
+
+    let resolved_root = resolve_project_root(Some(&tsconfig), &project_root, &files);
+    let resolved_tsconfig =
+        resolve_tsconfig_path(Some(&tsconfig), &project_root, &resolved_root, &files);
+
+    assert_eq!(resolved_root, project_root);
+    assert_eq!(resolved_tsconfig, Some(tsconfig));
+
+    let _ = std::fs::remove_dir_all(&resolved_root);
+}
+
+#[test]
+fn falls_back_to_cwd_resolution_when_files_have_no_tsconfig() {
+    let project_root = unique_case_dir("no-tsconfig");
+    let _ = std::fs::remove_dir_all(&project_root);
+    let src_dir = project_root.join("src");
+    std::fs::create_dir_all(&src_dir).unwrap();
+    let files = vec![src_dir.join("App.vue")];
+    std::fs::write(&files[0], "<template />").unwrap();
+
+    let resolved_root = resolve_project_root(None, &project_root, &files);
+    let resolved_tsconfig = resolve_tsconfig_path(None, &project_root, &resolved_root, &files);
+    let expected_root =
+        find_nearest_tsconfig_dir(&project_root).unwrap_or_else(|| project_root.clone());
+
+    assert_eq!(resolved_root, expected_root);
+    assert_eq!(
+        resolved_tsconfig,
+        resolved_root
+            .join("tsconfig.json")
+            .exists()
+            .then_some(resolved_root.join("tsconfig.json"))
+    );
+
+    let _ = std::fs::remove_dir_all(&project_root);
+}
+
+#[test]
+fn falls_back_to_common_file_parent_for_external_files_without_tsconfig() {
+    let case_root = std::env::temp_dir().join(format!(
+        "vize-check-runner-external-root-{}",
+        std::process::id()
+    ));
+    let _ = std::fs::remove_dir_all(&case_root);
+    let cwd = case_root.join("cwd");
+    let source_dir = case_root.join("external");
+    std::fs::create_dir_all(&cwd).unwrap();
+    std::fs::create_dir_all(&source_dir).unwrap();
+    let files = vec![source_dir.join("Repro.vue")];
+    std::fs::write(&files[0], "<template />").unwrap();
+
+    let resolved_root = resolve_project_root(None, &cwd, &files);
+    let resolved_tsconfig = resolve_tsconfig_path(None, &cwd, &resolved_root, &files);
+
+    assert_eq!(resolved_root, source_dir);
+    assert_eq!(resolved_tsconfig, None);
+
+    let _ = std::fs::remove_dir_all(&case_root);
+}
+
+#[test]
+fn resolve_declaration_dir_defaults_to_dist_types() {
+    let project_root = PathBuf::from("/workspace/project");
+    let tsconfig_options = TsconfigDeclarationOptions::default();
+    assert_eq!(
+        resolve_declaration_dir(None, &tsconfig_options, &project_root),
+        project_root.join("dist").join("types")
+    );
+    assert_eq!(
+        resolve_declaration_dir(Some(Path::new("types")), &tsconfig_options, &project_root),
+        project_root.join("types")
+    );
+}
+
+#[test]
+fn resolve_declaration_dir_uses_tsconfig_when_cli_dir_is_absent() {
+    let project_root = PathBuf::from("/workspace/project");
+    let tsconfig_options = TsconfigDeclarationOptions {
+        declaration_dir: Some(project_root.join("types")),
+        out_dir: Some(project_root.join("dist")),
+        declaration_map: Some(true),
+    };
+
+    assert_eq!(
+        resolve_declaration_dir(None, &tsconfig_options, &project_root),
+        project_root.join("types")
+    );
+    assert_eq!(
+        resolve_declaration_dir(Some(Path::new("custom")), &tsconfig_options, &project_root),
+        project_root.join("custom")
+    );
+
+    let out_dir_only = TsconfigDeclarationOptions {
+        declaration_dir: None,
+        out_dir: Some(project_root.join("dist")),
+        declaration_map: None,
+    };
+    assert_eq!(
+        resolve_declaration_dir(None, &out_dir_only, &project_root),
+        project_root.join("dist")
+    );
+}
+
+#[test]
+fn resolve_declaration_emit_options_uses_tsconfig_declaration_map() {
+    let project_root = unique_case_dir("declaration-options");
+    let _ = std::fs::remove_dir_all(&project_root);
+    std::fs::create_dir_all(&project_root).unwrap();
+    std::fs::write(
+        project_root.join("tsconfig.json"),
+        r#"{
+  "compilerOptions": {
+    "declarationDir": "types",
+    "declarationMap": true
+  }
+}"#,
+    )
+    .unwrap();
+
+    let options = resolve_declaration_emit_options(
+        None,
+        Some(&project_root.join("tsconfig.json")),
+        &project_root,
+    );
+
+    assert_eq!(options.out_dir, project_root.join("types"));
+    assert!(options.declaration_map);
+
+    let _ = std::fs::remove_dir_all(&project_root);
+}
+
+#[test]
+fn writes_nuxt_fallback_tsconfig_without_overwriting_existing_paths() {
+    let project_root = unique_case_dir("nuxt-fallback-tsconfig");
+    let _ = std::fs::remove_dir_all(&project_root);
+    std::fs::create_dir_all(&project_root).unwrap();
+    let base_tsconfig = project_root.join("tsconfig.base.json");
+    let tsconfig = project_root.join("tsconfig.json");
+    std::fs::write(
+        &base_tsconfig,
+        r##"{
+  "compilerOptions": {
+    "paths": {
+      "#base/*": ["types/*"]
+    }
+  }
+}"##,
+    )
+    .unwrap();
+    std::fs::write(
+        &tsconfig,
+        r##"{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "paths": {
+      "~/*": ["custom/*"]
+    }
+  }
+}"##,
+    )
+    .unwrap();
+
+    let wrapper = write_nuxt_fallback_tsconfig(
+        Some(&tsconfig),
+        &project_root,
+        &[
+            NuxtPathAlias {
+                pattern: "~/*".into(),
+                targets: vec!["app/*".into()],
+            },
+            NuxtPathAlias {
+                pattern: "#shared/*".into(),
+                targets: vec!["shared/*".into()],
+            },
+        ],
+    )
+    .unwrap();
+
+    let value: serde_json::Value =
+        serde_json::from_str(&std::fs::read_to_string(wrapper).unwrap()).unwrap();
+    let paths = value["compilerOptions"]["paths"].as_object().unwrap();
+    assert_eq!(value["extends"], tsconfig.to_string_lossy().as_ref());
+    assert_eq!(paths["#base/*"], serde_json::json!(["../../../types/*"]));
+    assert_eq!(paths["~/*"], serde_json::json!(["../../../custom/*"]));
+    assert_eq!(paths["#shared/*"], serde_json::json!(["../../../shared/*"]));
+
+    let _ = std::fs::remove_dir_all(&project_root);
+}
+
+#[test]
+fn validates_unsupported_corsa_server_counts() {
+    assert!(validate_corsa_server_count(None).is_ok());
+    assert!(validate_corsa_server_count(Some(1)).is_ok());
+    assert!(validate_corsa_server_count(Some(0)).is_err());
+    assert!(validate_corsa_server_count(Some(2)).is_err());
+}
